feature(refactoring): create update view handler
